requirements.txt
alembic==1.18.4
altair==6.0.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
argcomplete==3.6.3
attrs==25.4.0
bcrypt==5.0.0
blinker==1.9.0
build==1.4.0
CacheControl==0.14.4
cachetools==6.2.6
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
cleo==2.1.0
click==8.3.1
colorama==0.4.6
crashtest==0.4.1
cryptography==46.0.5
distlib==0.4.0
dnspython==2.8.0
dulwich==1.1.0
ecdsa==0.19.1
email-validator==2.3.0
fastapi==0.135.0
fastapi-cli==0.0.24
fastapi-cloud-cli==0.14.0
fastar==0.8.0
fastjsonschema==2.21.2
filelock==3.25.0
findpython==0.7.1
gitdb==4.0.12
GitPython==3.1.46
greenlet==3.3.2
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
idna==3.11
installer==0.7.0
jaraco.classes==3.4.0
jaraco.context==6.1.0
jaraco.functools==4.4.0
Jinja2==3.1.6
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
keyring==25.7.0
Mako==1.3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
more-itertools==10.8.0
msgpack==1.1.2
narwhals==2.15.0
numpy==2.4.1
packaging==26.0
pandas==2.3.3
passlib==1.7.4
pbs-installer==2026.3.3
pillow==12.1.0
pipx==1.8.0
pkginfo==1.12.1.2
platformdirs==4.9.2
poetry==2.3.2
poetry-core==2.3.1
protobuf==6.33.4
psycopg2-binary==2.9.11
pyarrow==23.0.0
pyasn1==0.6.2
pycparser==3.0
pydantic==2.12.5
pydantic-extra-types==2.11.0
pydantic-settings==2.13.1
pydantic_core==2.41.5
pydeck==0.9.1
Pygments==2.19.2
pymongo==4.16.0
PyMySQL==1.1.2
pyproject_hooks==1.2.0
python-dateutil==2.9.0.post0
python-discovery==1.1.0
python-dotenv==1.2.2
python-jose==3.5.0
python-multipart==0.0.22
pytz==2025.2
pywin32-ctypes==0.2.3
PyYAML==6.0.3
RapidFuzz==3.14.3
referencing==0.37.0
requests==2.32.5
requests-toolbelt==1.0.0
rich==14.3.3
rich-toolkit==0.19.7
rignore==0.7.6
rpds-py==0.30.0
rsa==4.9.1
sentry-sdk==2.54.0
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
SQLAlchemy==2.0.47
starlette==0.52.1
streamlit==1.53.1
tenacity==9.1.2
toml==0.10.2
tomlkit==0.14.0
tornado==6.5.4
trove-classifiers==2026.1.14.14
typer==0.24.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
userpath==1.9.2
uvicorn==0.41.0
virtualenv==21.1.0
watchdog==6.0.0
watchfiles==1.1.1
websockets==16.0
zstandard==0.25.0
redis==7.3.0
